other changes

+++ src/Demo.java
@@ -1,13 +1,13 @@
 public class Demo {
-    void show(int a, long b){
+    void show(int a, double b){
         System.out.println("short");
     }
-    void show(long a,int b){
+    void show(double a,int b){
         System.out.println("int");
     }
     public static void main(String[] args) {
         Demo o=new Demo();
-        o.show(4,5l);
+        o.show(4d,5);
 //
 
     }
